split results.dart into multiple files
